refactor(types): extract attachment tests to sibling file
